jconf.cpp: use erase from json object instead from the iterator
Instead of using the erase method from the iterator API, we should
use the one from the json object, which can save us the try-catch
block, the ugly Yoda comparision and the check if the key is the
one we want to remove.

In addition, we should remove the element from the high-level
hierarchy (if exists) before entering the loop in order to avoid
a situation in which we change the object while iterating it, that
can possibly lead to a segmentation fault.

Signed-off-by: Asaf Kahlon <asafka7@gmail.com>